io: Reduce system calls in common readall case
I came across the redundant fstat() and seemingly extra seek calls when
working on a tool that read lots of small files in directories as
config.

```python
from pathlib import Path

conf_dir = Path('.')
for file in conf_dir.glob('*.yaml'):
    contents = file.read_all()
    # do processing of contents
```

When fileio opens a file, fstat is always called. In the readall case,
it was always called again in addition to a seek. When scanning lots of
small files in directories that's a lot of redundant system calls. If
the file is big, we still save one fstat() (and we )

This saves work by
1. Caching the fstat() result so we only have to call fstat once. If the
file changes between open() and read() without a size, this _will_ mean
we have a less good estimate, but that was a matter of bad timing and we
will still read all the file in that case (we just get more loops).
2. If the file is small, don't bother with the lseek() (as well as
work to allow threads and suppress Windows IPH), and instead just use
the default buffer size.

At the end of the loop, the PyBytes is always resized to how much data
was read.

TODO:
Questions:
    Should we avoid fstat() up front altogether, and just start
        with SMALLCHUNK then fstat() and do the extra read?

TODO: In a folder with 50 .yaml files < 50 bytes, this reduces
TODO: Is there a good way in CPython to write a good test around this?0

TODO: This is part of a string I also have been pulling for:
1. Experiment with implementing the io module on top of io_uring
2. `python simple.py` reads the whole contents of the file 4 times, stats
the file at least 4 times, and does 11 seeks. That seems a little
excessive for simple script running... The OS should cache, but less work
is less work. Unfortunately that seems to be a mix of the parser and
import system, which is a ball I still haven't gotten my head around.
3. How much should we worry about excess reading if the file has been pre-read a bit?
4. Would it be better to use readv() for this case and get a single system call?
    -> Currently no: _Py_read does a lot, good to defer to
5. _blksize is only used for "buffered" reading, could we just call stat
   for it in that one case and save a member?